Refactor CandidateProfileForm component and isolated form component
