+++ src/components/TextArea/TextArea.component.jsx
@@ -1,0 +1,18 @@
+import {Pane, Label, Textarea } from 'evergreen-ui'
+
+const TextAreaAdapter = ({ input, ...rest }) => {
+    return (
+        <Pane display="flex" flexDirection="column" marginY="1em">
+            <Label htmlFor={input.name} marginY="0.5em" >
+                {rest.label}
+            </Label>
+            <Textarea {...input} {...rest}
+                id={input.name}
+                placeholder={rest.placeholder}
+                onChange={(event) => input.onChange(event)}
+            />
+        </Pane>
+    )
+}
+
+export default TextAreaAdapter;
